Update s3 support, moving it into the storage object.

+++ scripts/ZoneMinder/lib/ZoneMinder/Event.pm
@@ -463,33 +463,34 @@ sub delete_files {
       }
     } # end if Scheme eq Deep
 
-    # Now check for empty directories and delete them.
-    my @path_parts = split('/', $event_path);
-    pop @path_parts;
-    # Guaranteed the first part is the monitor id
-    Debug("Initial path_parts: @path_parts");
-    while ( @path_parts > 1 ) {
-      my $path = join('/', $storage_path, @path_parts);
-      my $dh;
-      if ( !opendir($dh, $path) ) {
-        Warning("Fail to open $path");
-        last;
-      }
-      my @dir =  readdir($dh);
-      closedir($dh);
-      if ( scalar(grep { $_ ne '.' and $_ ne '..' } @dir) == 0 ) {
-        Debug("Removing empty dir at $path");
-        if ( !rmdir $path ) {
-          Warning("Fail to rmdir $path: $!");
+    if ($Storage->Type() ne 's3fs') {
+      # Now check for empty directories and delete them.
+      my @path_parts = split('/', $event_path);
+      pop @path_parts;
+      # Guaranteed the first part is the monitor id
+      Debug("Initial path_parts: @path_parts");
+      while ( @path_parts > 1 ) {
+        my $path = join('/', $storage_path, @path_parts);
+        my $dh;
+        if ( !opendir($dh, $path) ) {
+          Warning("Fail to open $path");
           last;
         }
-      } else {
-        Debug(4, "Dir $path is not empty @dir");
-        last;
-      }
-      pop @path_parts;
-    } # end while path_parts
-
+        my @dir =  readdir($dh);
+        closedir($dh);
+        if ( scalar(grep { $_ ne '.' and $_ ne '..' } @dir) == 0 ) {
+          Debug("Removing empty dir at $path");
+          if ( !rmdir $path ) {
+            Warning("Fail to rmdir $path: $!");
+            last;
+          }
+        } else {
+          Debug(4, "Dir $path is not empty @dir");
+          last;
+        }
+        pop @path_parts;
+      } # end while path_parts
+    } # end if not s3fs
   } # end foreach Storage
 } # end sub delete_files
